refactor(tui): split ConversationPane into focused modules
`ConversationPane.tsx` had grown to 500 LOC mixing four concerns: the
`<Static>` scrollback transcript, per-entry renderers, viewport math,
and the live pane itself. Split along those seams so each file owns
one responsibility and ConversationPane.tsx shrinks to ~95 LOC of
glue:

  - `StaticConversationTranscript.tsx` — session header, scrollback
    item bookkeeping, and the `<Static>` wrapper.
  - `TranscriptEntry.tsx` — `TranscriptEntry`, `BoundedTranscriptEntry`,
    `LiveTranscriptEntry`, `ProcessEntryRow`.
  - `transcriptViewport.ts` — pure row-estimation and
    `selectPendingEntriesForViewport`.

Also drops the trivial `selectPendingEntriesForViewport` →
`selectLiveEntriesForViewport` identity wrapper (per CLAUDE.md
"Discouraged Factory Patterns") and the redundant double-nested
`<Box height={pendingRows}>` in the live pane.

`App.tsx` imports `StaticConversationTranscript` from its new module
directly. `ConversationPane.tsx` keeps re-exports for the symbols
test-kernel suites and downstream callers already import, so the
public surface is unchanged.
